Refactoring: grails.plugin.cookie -> grails.plugins.cookie (for consistency with Grails 3.x). Cleanup.
