+++ .travis.yml
@@ -1,6 +1,7 @@
 language: groovy
 jdk:
- - openjdk7
+    - openjdk7
+    - openjdk7
 before_script:
     - chmod +x grailsw
