+++ build.gradle
@@ -12,22 +12,22 @@ buildscript {
 }
 
 plugins {
-    id "io.spring.dependency-management" version "0.5.1.RELEASE"
-    id "com.jfrog.bintray" version "1.1"
+    id 'io.spring.dependency-management' version '0.5.1.RELEASE'
+    id "com.jfrog.bintray" version '1.1'
 }
 
-version "2.0-SNAPSHOT"
+version '2.0-SNAPSHOT'
 group "org.grails.plugins"
 
 apply plugin: 'maven-publish'
 apply plugin: 'eclipse'
 apply plugin: 'idea'
-apply plugin: "spring-boot"
-apply plugin: "org.grails.grails-plugin"
-apply plugin: "org.grails.grails-gsp"
+apply plugin: 'spring-boot'
+apply plugin: 'org.grails.grails-plugin'
+apply plugin: 'org.grails.grails-gsp'
 // Used for publishing to central repository, remove if not needed
-apply from:'https://raw.githubusercontent.com/grails/grails-profile-repository/master/profiles/plugin/templates/grailsCentralPublishing.gradle'
-apply from:'https://raw.githubusercontent.com/grails/grails-profile-repository/master/profiles/plugin/templates/bintrayPublishing.gradle'
+apply from: 'https://raw.githubusercontent.com/grails/grails-profile-repository/master/profiles/plugin/templates/grailsCentralPublishing.gradle'
+apply from: 'https://raw.githubusercontent.com/grails/grails-profile-repository/master/profiles/plugin/templates/bintrayPublishing.gradle'
 
 ext {
     grailsVersion = project.grailsVersion
@@ -37,11 +37,10 @@ ext {
 sourceCompatibility = 1.7
 targetCompatibility = 1.7
 
-
 repositories {
     mavenLocal()
     mavenCentral()
-    maven { url "https://repo.grails.org/grails/core" }
+    maven { url 'https://repo.grails.org/grails/core' }
 }
 
 dependencyManagement {
@@ -53,17 +52,17 @@ dependencyManagement {
 
 dependencies {
     provided 'org.springframework.boot:spring-boot-starter-logging'
-    provided "org.springframework.boot:spring-boot-starter-actuator"
-    provided "org.springframework.boot:spring-boot-autoconfigure"
-    provided "org.springframework.boot:spring-boot-starter-tomcat"
+    provided 'org.springframework.boot:spring-boot-starter-actuator'
+    provided 'org.springframework.boot:spring-boot-autoconfigure'
+    provided 'org.springframework.boot:spring-boot-starter-tomcat'
 
-    provided "org.grails:grails-web-boot"
-    provided "org.grails:grails-dependencies"
+    provided 'org.grails:grails-web-boot'
+    provided 'org.grails:grails-dependencies'
     provided 'javax.servlet:javax.servlet-api:3.1.0'
 
-    testCompile "org.grails:grails-plugin-testing"
+    testCompile 'org.grails:grails-plugin-testing'
 
-    console "org.grails:grails-console"
+    console 'org.grails:grails-console'
 }
 
 task wrapper(type: Wrapper) {
