--- grails-app/init/grails/cookie/Application.groovy
@@ -1,10 +1,0 @@
-package grails.cookie
-
-import grails.boot.GrailsApp
-import grails.boot.config.GrailsAutoConfiguration
-
-class Application extends GrailsAutoConfiguration {
-    static void main(String[] args) {
-        GrailsApp.run(Application, args)
-    }
-}
