+++ src/main/groovy/grails/plugins/cookie/CookieHelper.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package grails.plugin.cookie
+package grails.plugins.cookie
 
 import grails.web.api.ServletAttributes
 import grails.web.api.WebAttributes
