+++ src/main/groovy/grails/plugins/cookie/RequestResponseCookieExtension.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package grails.plugin.cookie
+package grails.plugins.cookie
 
 import groovy.util.logging.Commons
