+++ src/main/resources/META-INF/services/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,3 +1,3 @@
 moduleName=grails-cookie
 moduleVersion=2.0
-extensionClasses=grails.plugin.cookie.RequestResponseCookieExtension
+extensionClasses=grails.plugins.cookie.RequestResponseCookieExtension
