--- src/test/groovy/grails/plugin/cookie/CookieServiceRequestSpec.groovy
@@ -1,11 +1,0 @@
-package grails.plugin.cookie
-
-import grails.test.mixin.TestFor
-
-@TestFor(CookieService)
-class CookieServiceRequestSpec extends CookieRequestSpec {
-    @Override
-    def setup() {
-        obj = service
-    }
-}
