--- src/test/groovy/grails/plugin/cookie/CookieServiceResponseSpec.groovy
@@ -1,11 +1,0 @@
-package grails.plugin.cookie
-
-import grails.test.mixin.TestFor
-
-@TestFor(CookieService)
-class CookieServiceResponseSpec extends CookieResponseSpec {
-    @Override
-    def setup() {
-        obj = service
-    }
-}
