--- src/test/groovy/grails/plugin/cookie/RequestSpec.groovy
@@ -1,12 +1,0 @@
-package grails.plugin.cookie
-
-import grails.test.mixin.TestFor
-
-@TestFor(CookieService)
-class RequestSpec extends CookieRequestSpec {
-
-    @Override
-    def setup() {
-        obj = request
-    }
-}
