--- src/test/groovy/grails/plugin/cookie/ResponseSpec.groovy
@@ -1,12 +1,0 @@
-package grails.plugin.cookie
-
-import grails.test.mixin.TestFor
-
-@TestFor(CookieService)
-class ResponseSpec extends CookieResponseSpec {
-
-    @Override
-    def setup() {
-        obj = response
-    }
-}
